+++ scripts/test_gpio22.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 set -eu
 
-[ ! -e /dev/myled0 ] && echo "install myled driver!" && exit 1
+[ ! -e /dev/myled0 ] && { echo "install myled driver!"; exit 1; }
 
 for i in 1 0 1 0; do 
 	echo $i > /dev/myled0
